requirements.txt
pymysql~=1.0.2
paho-mqtt~=1.6.1
requests~=2.28.2
flask~=2.2.2
opencv-python~=4.7.0.68
python-dotenv~=0.21.1
pathlib~=1.0.1
wheel==0.38.4
setuptools==67.1.0
flask~=2.2.2
numpy~=1.24.1
imutils~=0.5.4